% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideoarchivedmedia_operations.R
\name{get_media_for_fragment_list}
\alias{get_media_for_fragment_list}
\title{Gets media for a list of fragments (specified by fragment number) from the archived data in an Amazon Kinesis video stream}
\usage{
get_media_for_fragment_list(StreamName, Fragments)
}
\arguments{
\item{StreamName}{[required] The name of the stream from which to retrieve fragment media.}

\item{Fragments}{[required] A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.}
}
\description{
Gets media for a list of fragments (specified by fragment number) from the archived data in an Amazon Kinesis video stream.
}
\details{
The following limits apply when using the \code{GetMediaForFragmentList} API:
\itemize{
\item A client can call \code{GetMediaForFragmentList} up to five times per second per stream.
\item Kinesis Video Streams sends media data at a rate of up to 25 megabytes per second (or 200 megabits per second) during a \code{GetMediaForFragmentList} session.
}
}
\section{Accepted Parameters}{
\preformatted{get_media_for_fragment_list(
  StreamName = "string",
  Fragments = list(
    "string"
  )
)
}
}

