% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideoarchivedmedia_package.R
\docType{package}
\name{paws.kinesisvideoarchivedmedia-package}
\alias{paws.kinesisvideoarchivedmedia}
\alias{paws.kinesisvideoarchivedmedia-package}
\title{paws.kinesisvideoarchivedmedia: Amazon Kinesis Video Streams Archived Media}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
